src/requirements.txt
Django==2.0.2
djangorestframework==3.8.2
flake8==3.5.0
Markdown==3.0.1
mccabe==0.6.1
pycodestyle==2.3.1
pydocstyle==2.1.1
pyflakes==1.6.0
pytz==2018.5
six==1.11.0
snowballstemmer==1.2.1